Move enums to domain
