+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/domain/enums/CategoryOfLaw.java
@@ -1,0 +1,6 @@
+package uk.gov.justice.laa.dstew.access.domain.enums;
+
+/** Domain enum representing the category of law for a legal aid application. */
+public enum CategoryOfLaw {
+  FAMILY
+}
