+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/domain/enums/MatterType.java
@@ -1,0 +1,6 @@
+package uk.gov.justice.laa.dstew.access.domain.enums;
+
+/** Domain enum representing the matter type for a legal aid application. */
+public enum MatterType {
+  SPECIAL_CHILDREN_ACT
+}
