--- data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/usecase/shared/parser/CategoryOfLaw.java
@@ -1,6 +1,0 @@
-package uk.gov.justice.laa.dstew.access.usecase.shared.parser;
-
-/** Usecase-level enum representing the category of law for a legal aid application. */
-public enum CategoryOfLaw {
-  FAMILY
-}
