--- data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/usecase/shared/parser/MatterType.java
@@ -1,6 +1,0 @@
-package uk.gov.justice.laa.dstew.access.usecase.shared.parser;
-
-/** Usecase-level enum representing the matter type for a legal aid application. */
-public enum MatterType {
-  SPECIAL_CHILDREN_ACT
-}
