+++ data-access-service/src/test/java/uk/gov/justice/laa/dstew/access/usecase/createapplication/CreateApplicationDomainMapperTest.java
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.Test;
 import uk.gov.justice.laa.dstew.access.domain.ApplicationDomain;
 import uk.gov.justice.laa.dstew.access.domain.IndividualDomain;
 import uk.gov.justice.laa.dstew.access.domain.ProceedingDomain;
-import uk.gov.justice.laa.dstew.access.usecase.shared.parser.CategoryOfLaw;
-import uk.gov.justice.laa.dstew.access.usecase.shared.parser.MatterType;
+import uk.gov.justice.laa.dstew.access.domain.enums.CategoryOfLaw;
+import uk.gov.justice.laa.dstew.access.domain.enums.MatterType;
 import uk.gov.justice.laa.dstew.access.usecase.shared.parser.ParsedAppContentDetails;
 import uk.gov.justice.laa.dstew.access.usecase.shared.parser.Proceeding;
 import uk.gov.justice.laa.dstew.access.utils.generator.DataGenerator;
